+++ src/MusicalScore/Graphical/VexFlow/VexFlowOctaveShift.ts
@@ -82,6 +82,14 @@ export class VexFlowOctaveShift extends GraphicalOctaveShift {
             const vve: VexFlowVoiceEntry = (gve as VexFlowVoiceEntry);
             if (vve?.vfStaveNote) {
                 this.endNote = vve.vfStaveNote;
+                this.endMeasure = graphicalStaffEntry.parentMeasure;
+                if (this.endMeasure?.parentSourceMeasure.Rules.OctaveShiftOnWholeMeasureNoteUntilEndOfMeasure &&
+                    vve.notes[0].sourceNote.isWholeMeasureNote()) {
+                    // draw whole note octave shift until end of measure
+                    //   Instead, we could try to fix the display of very short octaveshift brackets,
+                    //   which seem to overlap text (-> VF.TextBracket VexFlowPatch?).
+                    this.graphicalEndAtMeasureEnd = true;
+                }
                 return true;
             }
         }